+++ lat/legacy/generate_steering_vectors.py
